restructure test functions
