+++ vcd/common_test.go
@@ -1,0 +1,26 @@
+//go:build network || nsxt || gateway || ALL || functional
+// +build network nsxt gateway ALL functional
+
+package vcd
+
+import "testing"
+
+// updateEdgeGatewayTier0Dedication updates default NSX-T Edge Gateway used for tests to use Tier-0
+// gateway as dedicated for the purpose of some tests
+func updateEdgeGatewayTier0Dedication(t *testing.T, dedicatedTier0 bool) {
+	vcdClient := createSystemTemporaryVCDConnection()
+	org, err := vcdClient.GetOrgByName(testConfig.VCD.Org)
+	if err != nil {
+		t.Fatalf("error retrieving Org '%s': %s", testConfig.VCD.Org, err)
+	}
+	edge, err := org.GetNsxtEdgeGatewayByName(testConfig.Nsxt.EdgeGateway)
+	if err != nil {
+		t.Fatalf("error retrieving NSX-T Edge Gateway '%s': %s", testConfig.Nsxt.EdgeGateway, err)
+	}
+
+	edge.EdgeGateway.EdgeGatewayUplinks[0].Dedicated = dedicatedTier0
+	_, err = edge.Update(edge.EdgeGateway)
+	if err != nil {
+		t.Fatalf("error updating NSX-T Edge Gateway dedicated Tier 0 gateway usage to '%t': %s", dedicatedTier0, err)
+	}
+}
