+++ vcd/resource_vcd_nsxt_edgegateway_bgp_configuration_test.go
@@ -12,24 +12,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
-func updateEdgeGatewayTier0Dedication(t *testing.T, dedicatedTier0 bool) {
-	vcdClient := createSystemTemporaryVCDConnection()
-	org, err := vcdClient.GetOrgByName(testConfig.VCD.Org)
-	if err != nil {
-		t.Fatalf("error retrieving Org '%s': %s", testConfig.VCD.Org, err)
-	}
-	edge, err := org.GetNsxtEdgeGatewayByName(testConfig.Nsxt.EdgeGateway)
-	if err != nil {
-		t.Fatalf("error retrieving NSX-T Edge Gateway '%s': %s", testConfig.Nsxt.EdgeGateway, err)
-	}
-
-	edge.EdgeGateway.EdgeGatewayUplinks[0].Dedicated = dedicatedTier0
-	_, err = edge.Update(edge.EdgeGateway)
-	if err != nil {
-		t.Fatalf("error updating NSX-T Edge Gateway dedicated Tier 0 gateway usage to '%t': %s", dedicatedTier0, err)
-	}
-}
-
 // TestAccVcdNsxtEdgeBgpConfigTier0 tests out NSX-T Edge Gateway BGP Configuration using dedicated
 // Tier-0 gateway
 func TestAccVcdNsxtEdgeBgpConfigTier0(t *testing.T) {
